Incorporate more state into the StreamHelper (removing the need for TrackToHostData).
Register the available helper into a concurrent queue (for now the one we pick requires inheritance from TObject). We, for now, just underfill it and execute the stream immediately (next is the delay of the execution until it is fuller). Add to the helper an array holding the 'id'/'reference' of the track. Continue to propagate particle that are 'still' alive. Remove the need for the 2nd duplicate code in the loop (replace by picking the next available Stream).