Reorganize Propagator initialization.
Some part of the initialization can be done before loading the geometry while
some other part (in particular the one depending on the max depth) must be done
afterwards.

So that the Geometry loading routine can used the workload manager (in a manner
consistent with the rest of the code) we split these two parts in two differnt
routines.

Also since the ordering the call is essential, the routines are made private.