GEANT-102  Get rid of long initialisation caused by checking the used logical volumes.
Temporary solution (only for the current cms2015.gdml) to skip checking of used
logical volumes that is necessary because of a bug in Geant4 GDML reader. This check
will be removed when Geant4-10.2.beta (with the GDML reader fix) will be available.
Scoring option as an input parameter was also added to testcomplex.