GEANT-103 Properly handle the end of simulation case.
Handle the case where the broker is the one to process the last basket, but did not
fill its own baskets (and thus did not schedule the kernel to run).  In that case
rather than waiting forever for a basket that will never come, force the launching
of the kernel.