Unify implementations of GUAliasSampler interfaces using VecCore
This eliminates template specializations for the Vc backend of
GUAliasSampler::GatherAlias() and GUAliasSampler::GetPDF() functions
in favor of generic implementations using VecCore's backend interface.

Note: It is not possible to use the Gather() interface at the moment,
because types being gathered are not compatible with the pointer being
passed to the GUAliasSampler::GatherAlias() function. In other words,
the table for indices fAlias is filled with 32bit ints, which in AVX
have 8 elements in SIMD types, while aliasInd, the variable being
gathered, is for doubles, so it only has 4 elements, and is represented
as an SSE vector of 32bit integers containing 4 elements.